Use http.Header methods instead of being clever
golang/go@1e814df79bdacc618dfe9768dfdece16c7b0a499 tried to avoid
unnecessary calls to CanonicalHeaderKey in the standard library, but it
is not clear that there is any meaningful performance gain, nor is it
used consistently. This change mostly reverts the manual manipulation
of the http.Header map by using the provided methods instead.